requirements.txt
torch>=1.6.0
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=7.2.0
opencv-python>=4.4.0
scikit-learn>=0.23.2
tqdm>=4.48.0
tensorboard>=2.3.0
